Rename test files and Move helper functions
